refactor: isolate trade params logic
